modules/base/scripts/voice/requirements.txt
openai-whisper
numpy
